+++ gitman/__main__.py
@@ -1,0 +1,7 @@
+"""Package entry point."""
+
+from .cli import main
+
+
+if __name__ == '__main__':  # pragma: no cover (manual test)
+    main()