+++ gitman/models/__init__.py
@@ -1,0 +1,4 @@
+"""Domain models for interacting with dependency configuration."""
+
+from .source import Source
+from .config import Config, load_config